Restructure cmds and main.go
