+++ cmd/evaluate.go
